Resume subMenuIcon
According APPLINK-15683 SDL must restore AddSubMenu with all parameters

Related to APPLINK-23379

+++ src/components/application_manager/src/resumption/resumption_data_db.cc
@@ -803,14 +803,26 @@ bool ResumptionDataDB::SelectSubmenuData(
   /* Position of data in "select_sub_menu" :
      field "menuID" from table "subMenu" = 0
      field "menuName" from table "subMenu" = 1
-     field "position" from table "subMenu" = 2*/
+     field "position" from table "subMenu" = 2
+     field "value" from table "image" = 3
+     field "imageType" from table "image" = 4*/
   uint32_t i = 0;
   while (select_sub_menu.Next()) {
     SmartObject array_item(SmartType_Map);
+    SmartObject submenu_icon(SmartType_Map);
     array_item[strings::menu_id] = select_sub_menu.GetInteger(0);
     array_item[strings::menu_name] = select_sub_menu.GetString(1);
     if (!(select_sub_menu.IsNull(2))) {
       array_item[strings::position] = select_sub_menu.GetInteger(2);
+    }
+    if (!(select_sub_menu.IsNull(3))) {
+      submenu_icon[strings::value] = select_sub_menu.GetString(3);
+    }
+    if (!(select_sub_menu.IsNull(4))) {
+      submenu_icon[strings::image_type] = select_sub_menu.GetInteger(4);
+    }
+    if (!submenu_icon.empty()) {
+      array_item[strings::sub_menu_icon] = submenu_icon;
     }
     saved_app[strings::application_submenus][i++] = array_item;
   }
@@ -1447,6 +1459,12 @@ bool ResumptionDataDB::DeleteSavedSubMenu(const std::string& policy_app_id,
                                           const std::string& device_id) {
   SDL_AUTO_TRACE();
 
+  if (!ExecQueryToDeleteData(
+          policy_app_id, device_id, kDeleteImageFromSubMenu)) {
+    SDL_WARN("Incorrect delete image from SubMenu.");
+    return false;
+  }
+
   if (!ExecQueryToDeleteData(policy_app_id, device_id, kDeleteSubMenu)) {
     SDL_WARN("Incorrect delete from subMenu.");
     return false;
@@ -1642,6 +1660,7 @@ bool ResumptionDataDB::ExecInsertImage(
     int64_t& image_primary_key, const smart_objects::SmartObject& image) const {
   SDL_AUTO_TRACE();
   using namespace app_mngr;
+
   utils::dbms::SQLQuery count_image_query(db());
   utils::dbms::SQLQuery query(db());
   uint16_t count_image = 0;
@@ -1991,6 +2010,7 @@ bool ResumptionDataDB::InsertSubMenuData(
   }
   utils::dbms::SQLQuery query_insert_submenu(db());
 
+  int64_t image_primary_key = 0;
   if (!query_insert_submenu.Prepare(kInsertToSubMenu)) {
     SDL_WARN("Problem with verification queries for insertion submenu");
     return false;
@@ -1998,11 +2018,25 @@ bool ResumptionDataDB::InsertSubMenuData(
   /* Positions of binding data for "query_insert_submenu":
      field "menuID" from table "submenu" = 0
      field "menuName" from table "submenu" = 1
-     field "position" from table "submenu" = 2*/
+     field "position" from table "submenu" = 2
+     field "idimage" from table "submenu" = 3
+*/
   for (size_t i = 0; i < length_submenu_array; ++i) {
     query_insert_submenu.Bind(0, (submenus[i][strings::menu_id]).asInt());
     query_insert_submenu.Bind(1, (submenus[i][strings::menu_name]).asString());
+
     CustomBind(strings::position, submenus[i], query_insert_submenu, 2);
+
+    if (submenus[i].keyExists(strings::sub_menu_icon)) {
+      if (!ExecInsertImage(image_primary_key,
+                           submenus[i][strings::sub_menu_icon])) {
+        SDL_WARN("Problem with insert submenu icon to DB");
+        return false;
+      }
+      query_insert_submenu.Bind(3, image_primary_key);
+    } else {
+      query_insert_submenu.Bind(3);
+    }
 
     if (!query_insert_submenu.Exec()) {
       SDL_WARN("Incorrect insertion of submenu data");
